requirements.txt
streamlit==1.27.0
pandas==2.0.3
pytesseract==0.3.10
pdfplumber==0.10.2
reportlab==3.6.12
pillow==9.5.0
numpy==1.26.0
google-generativeai==0.3.0

